:penguin: Fix rubocop error
